chore(): move rack-mini-profiler test cases to a separate test file
