+++ test/controllers/admin/comfy/v2/dashboard_controller_test.rb
@@ -16,12 +16,6 @@ class Comfy::Admin::V2::DashboardControllerTest < ActionDispatch::IntegrationTes
       slug: 'test-cms-page',
     )
     @api_resource = api_resources(:one)
-
-    @restarone_subdomain = Subdomain.find_by(name: 'restarone')
-    Apartment::Tenant.switch @restarone_subdomain.name do
-      @restarone_user = User.find_by(email: 'contact@restarone.com')
-      @restarone_user.update(can_manage_analytics: true)
-    end
   end
 
   test "should deny #v2_dashboard if not signed in" do
@@ -246,44 +240,4 @@ class Comfy::Admin::V2::DashboardControllerTest < ActionDispatch::IntegrationTes
 
     assert_response :success
   end
-
-  # Rack Mini-Profiler test cases - START 
-  test "#dashboard: should show mini-profiler badge if user is permissioned properly" do
-    Rails.env.stubs(:test?).returns(false)
-    Rack::MiniProfiler::ClientSettings.any_instance.stubs(:has_valid_cookie?).returns(true)
-
-    Apartment::Tenant.switch @restarone_subdomain.name do
-      @restarone_user.update(can_access_admin: true, show_profiler: true)
-
-      sign_in(@restarone_user)
-      get v2_dashboard_url(subdomain: @restarone_subdomain.name)
-
-      assert_response :success
-      assert response.headers.has_key?('X-MiniProfiler-Ids')
-    end
-  end
-
-  test "#dashboard: should not show mini-profiler badge if user cannot access admin or profiler" do
-    Rails.env.stubs(:test?).returns(false)
-    Rack::MiniProfiler::ClientSettings.any_instance.stubs(:has_valid_cookie?).returns(true)
-
-    Apartment::Tenant.switch @restarone_subdomain.name do
-      @restarone_user.update(can_access_admin: false, show_profiler: true)
-
-      sign_in(@restarone_user)
-      get v2_dashboard_url(subdomain: @restarone_subdomain.name)
-      assert_response :redirect
-
-      refute response.headers.has_key?('X-MiniProfiler-Ids')
-      
-      @restarone_user.update(can_access_admin: true, show_profiler: false)
-
-      sign_in(@restarone_user)
-      get v2_dashboard_url(subdomain: @restarone_subdomain.name)
-      assert_response :success
-
-      refute response.headers.has_key?('X-MiniProfiler-Ids')
-    end
-  end
-  # Rack Mini-Profiler test cases - END
 end
